api/requirements.txt
fastapi
uvicorn
python-multipart
minio
psycopg2-binary